Update LegrandSwitch.groovy to Handle Switches with null States
When a Legrand switch is initially imported into SmartThings for the first time, it may be in a "null" state. Prior to this edit, there was no way for a null-valued switch to trigger a sendEvent. Since a sendEvent is the only way to change the switch's state from "null" to "on" or "off," this inability to trigger a sendEvent left the null-valued switch with a "null" state in perpetuity. And while in a "null" state, the switch's status cannot be accurately reflected in SmartThings.