Update LegrandConnect.groovy to Accept Zone Property Changes When ZID = 0
Prior to this change, a status report with ZID = 0 would log an error instead of grabbing the status of the light with ZID = 0. This was due to an if statement searching for any negated reqJSON.ZID value (considering a value of 0 as a negation). This resulted in an undesired if statement being entered when a status was received from a light with ZID = 0.